requirements.txt
fastapi
uvicorn
motor==3.3.2
pymongo==4.6.1
python-dotenv
pydantic